cricketdata: International Cricket Data

Data on international and other major cricket matches from ESPNCricinfo https://www.espncricinfo.com and Cricsheet https://cricsheet.org. This package provides some functions to download the data into tibbles ready for analysis.

See also

Author

Maintainer: Rob Hyndman Rob.Hyndman@monash.edu

Authors:

Charles Gray C.Gray@latrobe.edu.au

Sayani Gupta Sayani.Gupta@monash.edu

Timothy Hyndman Timothy.Hyndman@gmail.com

Hassan Rafique dazzalytics@protonmail.com

Jacquie Tran jac@jacquietran.com

Other contributors:

Puwasala Gamakumara Puwasala.Gamakumara@monash.edu [contributor]

Alex Whan alexwhan@gmail.com [contributor]